first commit on etch a sketch
got the HTML layout sorted, contrasting colours in CSS for visibility
javascript is drawing the boxes fine for initial phase and filling
box + wrapping as expected. next need to add the mouseover event
to the boxes but having difficulty figuring out how to add this to
the nodelist, will review the mouseover event from the drum kit
tutorial.